Fix build and add CHANGELOG
