+++ src/pages-lib/[username].test.ts
@@ -1,7 +1,7 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import { lookupProfile } from 'blockstack';
 import * as Sentry from '@sentry/node';
-import PublicHomePage from './[username]';
+import { PublicHomePage } from './[username]';
 
 jest.mock('blockstack');
 jest.mock('@sentry/node');
