+++ src/pages-lib/[username]/[storyId].test.tsx
@@ -1,7 +1,7 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import { lookupProfile } from 'blockstack';
 import * as Sentry from '@sentry/node';
-import PublicStoryPage from './[storyId]';
+import { PublicStoryPage } from './[storyId]';
 
 jest.mock('blockstack');
 jest.mock('@sentry/node');
